Update user.rb
sorry, this is the follow on the update #5628. I removed the check for specific URL

previous end of the thread:

I am playing with a site on a raspberry pi (got the domain 18xxconductor.fun for cheap if you can believe that). I plan on basically just handling the JSON the site already sends to trigger other webhooks allowing different discord channels in this case to be alerted for different games. 

Thanks!